Disconnected service fails to check for pre-configured networks
Issues:
Fixes #516

Problem:
In order to support opflex networks on LBaaSv1 the network was pre-configured
on the BigIP and then the network was added to the common_network_ids config
item in the ini file. This doesn't work on the v2 version of the driver b/c
the is_service_connected method of disconnected_service only checks if the
network type is 'flat' or if the segment_id is defined. The opflex network
type in Neutron does not define a segment id and so the check fails, raising
an exception.

This exception also by-passed the lbaas cleanup code resulting in the need to
manually go into the neutron database to clear out loadbalancers in 'ERROR' state.

Analysis:
Preform an extra check when determining if a service is connected to verify
that the network is not in the list of preprovisioned networks.

Tests: