Common network bound to disconnected_network for opflex deployment
Issues:
Fixes #538,#523

Problem:
Virtual server is bound to the dummy network disconnected_network
when the segmentation ID is not present.  This occurs even if the
network is defined in the list of common_network_ids.

When a listener is created on a network that does not have a seg
id, the virtual server will be created on the disconnected net,
but it should be connected on the vlan paired with the network
id.

Analysis:
Create a new function in the service_adapter to return true if
the VIP is associated with a network defined in the list of
common networks.  Use this function when creating a virtual server
and determining whether or not a service is connected.

Also, extend the function to add a common_network to the list
of networks that a virtual is enabled on.

Tests:
Disconnected network tests
Added test_featureoff_nosegid_create_listener_common_lb_net