pare down lines of code by rming cert_manager
Issues:
Fixes #542

Problem: The agent had code in place in case of future use.
We intend to pursue a policy of Late Abstraction,
We implement, then we reimplement, then we recognize the
commonalities and implement the simplifying abstraction.

Analysis:  This commit removes the cert_manager module, and
updates imports.  In order to fulfill the expose-to external we
update the required method docstrings.

Tests:  N/A