Fix requirements and systest scripts to support newton
Issues:
Fixes #670

Problem:
Several things need to be updated to support the newton release.
First, we should fix the requirements files to support newton
requirements. The docs should be modified to reflect the new branch.

Analysis:
Updated the doc references to openstack releases. Updated the functest
and unittest requirements files to reflect the new branch.

Tests:
Testing build in travis on my fork before creating the PR. If the build
passes, then the changes have worked.